SW-6029 Allow phase 0 scores for applications
Allow clients to submit phase 0 scores for projects that aren't in cohorts but
have applications.

Since this requires mapping application statuses to phases, move the logic for
that mapping into `PhaseChecker` (which is now renamed to `ProjectCohortFetcher`)
so there's a single source of truth for it, rather than also having it in
`ProjectStore`.
